codec: re-cut identity/event + object model to the reference model
Land the geometry-free half of the 0.2.3 reference-model re-cut, browser-verified
end to end (npc → edge → shard → worker → state → pixijs; wolves + terrain render).

Hot / identity / event side:
- refs.rs: server_reference = realm_id:8|server_id:8 (geographic, #10);
  entity_reference = reserved:10|reference_id:6|server_reference:16|object_reference:32
  (#5), hot = REF_HOT|server_reference|hot_reference:32; dropped the dead functional
  server_type/action_reference/zone_reference machinery (0 live uses). Cold/positional
  kept as a REF_COLD variant (interim world-global zone_id|location|layer).
- event_reference u64 -> u32: event_log PK + holder/open_rows + every reducer sig +
  vm::await_gate/encode_await. entity_key kept u64 (holds the new-layout, server-
  qualified entity_reference — #5 without narrowing the PK).

Object-model side (object.rs -> definition / position / data):
- definition_reference:u32 = type_id:4|subtype_id:12|kind_id:12|variant_id:4 (drop
  subkind, kind 10->12b); type_reference/kind_reference halves.
- cold entry kind_pos_reference:u32 = kind_reference:16|tile:8|data:8; data:8 =
  sub_position:3|rotation:2|aux:3.
- geographic cold_reference:u32 = region:8|zone:8|tile:8|layer:8 (realm rides
  server_reference); region_zone_reference naming (not macro_*).

Propagated through the pipeline macro, regenerated ST bindings (edge/worker/master),
worker, edge, tick vm, client/core, npc, wasm, pixijs (MoverLayer filters REF_HOT),
and worldgen (pack_cold_entry). The cold table stays on its zone_id key for now — the
legacy world-global zone_id retirement (drop surface, region_zone:u16 key) is next.

+++ server/edge/src/bindings/shard/event_log_table.rs
@@ -104,7 +104,7 @@ impl<'ctx> __sdk::TableWithPrimaryKey for EventLogTableHandle<'ctx> {
         /// but to directly chain method calls,
         /// like `ctx.db.event_log().event_reference().find(...)`.
         pub struct EventLogEventReferenceUnique<'ctx> {
-            imp: __sdk::UniqueConstraintHandle<EventLog, u64>,
+            imp: __sdk::UniqueConstraintHandle<EventLog, u32>,
             phantom: std::marker::PhantomData<&'ctx super::RemoteTables>,
         }
 
@@ -112,7 +112,7 @@ impl<'ctx> __sdk::TableWithPrimaryKey for EventLogTableHandle<'ctx> {
             /// Get a handle on the `event_reference` unique index on the table `event_log`.
             pub fn event_reference(&self) -> EventLogEventReferenceUnique<'ctx> {
                 EventLogEventReferenceUnique {
-                    imp: self.imp.get_unique_constraint::<u64>("event_reference"),
+                    imp: self.imp.get_unique_constraint::<u32>("event_reference"),
                     phantom: std::marker::PhantomData,
                 }
             }
@@ -121,7 +121,7 @@ impl<'ctx> __sdk::TableWithPrimaryKey for EventLogTableHandle<'ctx> {
         impl<'ctx> EventLogEventReferenceUnique<'ctx> {
             /// Find the subscribed row whose `event_reference` column value is equal to `col_val`,
             /// if such a row is present in the client cache.
-            pub fn find(&self, col_val: &u64) -> Option<EventLog> {
+            pub fn find(&self, col_val: &u32) -> Option<EventLog> {
                 self.imp.find(col_val)
             }
         }
@@ -130,7 +130,7 @@ impl<'ctx> __sdk::TableWithPrimaryKey for EventLogTableHandle<'ctx> {
 pub(super) fn register_table(client_cache: &mut __sdk::ClientCache<super::RemoteModule>) {
 
     let _table = client_cache.get_or_make_table::<EventLog>("event_log");
-    _table.add_unique_constraint::<u64>("event_reference", |row| &row.event_reference);
+    _table.add_unique_constraint::<u32>("event_reference", |row| &row.event_reference);
 }
 
 #[doc(hidden)]
